+++ src/utils/leb128.ts
@@ -6,16 +6,21 @@
  * @param buffer the buffer of bytes to read
  * @returns the decoded number
  */
-export function decode(buffer: Uint8Array): number {
+export function decode(buffer: Uint8Array): { result: number, tail: Uint8Array } {
   let result = 0
   let shift = 0
+  let bytesRead = 0
   for (const byte of buffer) {
+    bytesRead++;
     result |= (byte & 0x7f) << shift
     if ((0x80 & byte) === 0) break
     shift += 7
   }
 
-  return result
+  return {
+    result,
+    tail: Uint8Array.from(buffer.slice(bytesRead)),
+  }
 }
 
 /**
@@ -27,7 +32,7 @@ export function decode(buffer: Uint8Array): number {
 export function encode(value: number): Uint8Array {
   const result = [];
 
-  while (value !== 0) {
+  while (true) {
     let byte_ = value & 0x7f;
     value >>= 7;
     if (value !== 0) {
@@ -35,6 +40,10 @@ export function encode(value: number): Uint8Array {
     } 
 
     result.push(byte_);
+
+    if (value === 0) {
+      break;
+    }
   }
 
   return Uint8Array.from(result)